Various tidy-up

Internalised all the properties in the Resources structure, providing methods (well one) where they are needed externally

Tidied up the bind-mounts code

Tidied up the VPMem code for utility VMs. Still some more work needed.

Fixed the LCOW scratch path to under /run/gcs/c/n/scratch (rather than /run/gcs/c/n/upper)

Renamed several variables so that it's much clearer what they actually are. Code was extremely confusing (to me...)

Finished removing the aliasing of schema2 to hcsschemav2

Split the HCS document creation code in hcsoci into hcsdoc_wcow.go and hcsdoc_lcow.go for clarity.`

Got rid of a few TODOs

Fixed SCSI to pass the hosted settings through on remove correctly when mapped. Otherwise you have inconsistent "surprise" removal

Removed CreateProcess from internal/uvm and moved to lcow/process. Extended it to cope with both in UVM and for init process

Finally got an LCOW container init process to start in test code and get stdout back to prove a simple pod scenario

Updated rootfs2vhd to use lcow.CreateProcess instead of shelling out to hcsdiag exec (another TODO removed)
