+++ functional/utilities/requiresbuild.go
@@ -7,7 +7,7 @@ import (
 )
 
 func RequiresBuild(t *testing.T, b uint16) {
-	if osversion.GetOSVersion().Build < b {
+	if osversion.Get().Build < b {
 		t.Skipf("Requires build %d+", b)
 	}
 }
