+++ internal/lcow/process.go
@@ -1,13 +1,15 @@
-package uvm
+package lcow
 
 import (
 	"fmt"
 	"io"
 	"strings"
+	"time"
 
 	"github.com/Microsoft/hcsshim/internal/copywithtimeout"
 	"github.com/Microsoft/hcsshim/internal/hcs"
-	"github.com/Microsoft/hcsshim/internal/schema1"
+	"github.com/Microsoft/hcsshim/internal/schema2"
+	"github.com/Microsoft/hcsshim/internal/schemaversion"
 	specs "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/sirupsen/logrus"
 )
@@ -20,19 +22,27 @@ type ByteCounts struct {
 	Err int64
 }
 
-// ProcessOptions are the set of options which are passed to CreateProcess() to
+// ProcessOptions are the set of options which are passed to CreateProcessEx() to
 // create a utility vm.
 type ProcessOptions struct {
-	Process    *specs.Process
-	Stdin      io.Reader  // Optional reader for sending on to the processes stdin stream
-	Stdout     io.Writer  // Optional writer for returning the processes stdout stream
-	Stderr     io.Writer  // Optional writer for returning the processes stderr stream
-	ByteCounts ByteCounts // How much data to copy on each stream if they are supplied. 0 means to io.EOF.
+	HCSSystem         *hcs.System
+	Process           *specs.Process
+	Stdin             io.Reader     // Optional reader for sending on to the processes stdin stream
+	Stdout            io.Writer     // Optional writer for returning the processes stdout stream
+	Stderr            io.Writer     // Optional writer for returning the processes stderr stream
+	CopyTimeout       time.Duration // Timeout for the copy
+	CreateInUtilityVm bool          // If the compute system is a utility VM
+	ByteCounts        ByteCounts    // How much data to copy on each stream if they are supplied. 0 means to io.EOF.
 }
 
-// CreateProcess creates a process in the utility VM. This is only used
-// on LCOW to run processes for remote filesystem commands, utilities, and debugging.
-// It will return an error on a Windows utility VM.
+// CreateProcess creates a process either in an LCOW utility VM, or for starting
+// the init process. TODO: Potentially extend for exec'd processes.
+//
+// It's essentially a glorified wrapper around hcs.ComputeSystem CreateProcess used
+// for internal purposes.
+//
+// This is used on LCOW to run processes for remote filesystem commands, utilities,
+// and debugging.
 //
 // It optional performs IO copies with timeout between the pipes provided as input,
 // and the pipes in the process.
@@ -43,49 +53,56 @@ type ProcessOptions struct {
 // was sent/received from the process.
 //
 // It is the responsibility of the caller to call Close() on the process returned.
-//
-// TODO: This could be removed as on LCOW as we could run a privileged container.
 
-func (uvm *UtilityVM) CreateProcess(opts *ProcessOptions) (*hcs.Process, *ByteCounts, error) {
-	if uvm.hcsSystem == nil {
-		return nil, nil, fmt.Errorf("cannot CreateProcess - no hcsSystem handle")
-	}
-	if uvm.operatingSystem != "linux" {
-		return nil, nil, fmt.Errorf("CreateProcess only supported on linux utility VMs")
-	}
+func CreateProcess(opts *ProcessOptions) (*hcs.Process, *ByteCounts, error) {
 
-	if opts.Process == nil {
-		return nil, nil, fmt.Errorf("no Process passed to CreateProcessEx")
-	}
-
+	var environment = make(map[string]string)
 	copiedByteCounts := &ByteCounts{}
-	commandLine := strings.Join(opts.Process.Args, " ")
-	environment := make(map[string]string)
-	for _, v := range opts.Process.Env {
-		s := strings.SplitN(v, "=", 2)
-		if len(s) == 2 && len(s[1]) > 0 {
-			environment[s[0]] = s[1]
-		}
+
+	if opts == nil {
+		return nil, nil, fmt.Errorf("no options supplied")
 	}
 
-	if uvm.operatingSystem == "linux" {
+	if opts.HCSSystem == nil {
+		return nil, nil, fmt.Errorf("no HCS system supplied")
+	}
+
+	if opts.CreateInUtilityVm && opts.Process == nil {
+		return nil, nil, fmt.Errorf("process must be supplied for UVM process")
+	}
+
+	// Don't pass a process in if this is an LCOW container. This will start the init process.
+	if opts.Process != nil {
+		for _, v := range opts.Process.Env {
+			s := strings.SplitN(v, "=", 2)
+			if len(s) == 2 && len(s[1]) > 0 {
+				environment[s[0]] = s[1]
+			}
+		}
 		if _, ok := environment["PATH"]; !ok {
 			environment["PATH"] = "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:"
 		}
 	}
 
-	processConfig := &schema1.ProcessConfig{
-		EmulateConsole:    false,
+	processConfig := &schema2.ProcessConfig{
+		SchemaVersion:     schemaversion.SchemaV20(),
 		CreateStdInPipe:   (opts.Stdin != nil),
 		CreateStdOutPipe:  (opts.Stdout != nil),
 		CreateStdErrPipe:  (opts.Stderr != nil),
-		CreateInUtilityVm: true,
-		WorkingDirectory:  opts.Process.Cwd,
-		Environment:       environment,
-		CommandLine:       commandLine,
+		CreateInUtilityVm: opts.CreateInUtilityVm,
+		EmulateConsole:    false,
 	}
 
-	proc, err := uvm.hcsSystem.CreateProcess(processConfig)
+	if opts.Process != nil {
+		processConfig.Environment = environment
+		processConfig.CommandLine = strings.Join(opts.Process.Args, " ")
+		processConfig.WorkingDirectory = opts.Process.Cwd
+		if processConfig.WorkingDirectory == "" {
+			processConfig.WorkingDirectory = `/`
+		}
+	}
+
+	proc, err := opts.HCSSystem.CreateProcess(processConfig)
 	if err != nil {
 		logrus.Debugf("failed to create process: %s", err)
 		return nil, nil, err
@@ -103,8 +120,8 @@ func (uvm *UtilityVM) CreateProcess(opts *ProcessOptions) (*hcs.Process, *ByteCo
 		if copiedByteCounts.In, err = copywithtimeout.Copy(processStdin,
 			opts.Stdin,
 			opts.ByteCounts.In,
-			fmt.Sprintf("CreateProcessEx: to stdin of %q", commandLine),
-			defaultTimeoutSeconds); err != nil {
+			"stdin",
+			opts.CopyTimeout); err != nil {
 			return nil, nil, err
 		}
 
@@ -123,8 +140,8 @@ func (uvm *UtilityVM) CreateProcess(opts *ProcessOptions) (*hcs.Process, *ByteCo
 		if copiedByteCounts.Out, err = copywithtimeout.Copy(opts.Stdout,
 			processStdout,
 			opts.ByteCounts.Out,
-			fmt.Sprintf("CreateProcessEx: from stdout from %q", commandLine),
-			defaultTimeoutSeconds); err != nil {
+			"stdout",
+			opts.CopyTimeout); err != nil {
 			return nil, nil, err
 		}
 	}
@@ -135,12 +152,10 @@ func (uvm *UtilityVM) CreateProcess(opts *ProcessOptions) (*hcs.Process, *ByteCo
 		if copiedByteCounts.Err, err = copywithtimeout.Copy(opts.Stderr,
 			processStderr,
 			opts.ByteCounts.Err,
-			fmt.Sprintf("CreateProcessEx: from stderr of %s", commandLine),
-			defaultTimeoutSeconds); err != nil {
+			"stderr",
+			opts.CopyTimeout); err != nil {
 			return nil, nil, err
 		}
 	}
-
-	logrus.Debugf("hcsshim: CreateProcessEx success: %q", commandLine)
 	return proc, copiedByteCounts, nil
 }
