+++ internal/uvm/counter.go
@@ -7,8 +7,8 @@ var (
 	counter uint64
 )
 
-// ContainerCounter is used for LCOW. It's where we mount the overlay filesystem
-// inside the utility VM when mounting container layers. eg /tmp/cN
+// ContainerCounter is used for where we layout things for a container in
+// a utility VM. For WCOW it'll be C:\c\N\. For LCOW it'll be /run/gcs/c/N/.
 func (uvm *UtilityVM) ContainerCounter() uint64 {
 	m.Lock()
 	defer m.Unlock()
