+++ internal/uvm/system.go
@@ -1,0 +1,7 @@
+package uvm
+
+import "github.com/Microsoft/hcsshim/internal/hcs"
+
+func (uvm *UtilityVM) ComputeSystem() *hcs.System {
+	return uvm.hcsSystem
+}
